feat: adequação de mensagem para baseRepository

+++ Src/Modules/Message/Models/MessageModel.cs
@@ -6,16 +6,16 @@ public class MessageModel : BaseModel
 {
     public string Texto { get; set; } = string.Empty;
 
-    public int IdUsuario { get; set; }
-    public int IdTicket { get; set; }
+    public long IdUsuario { get; set; }
+    public long IdTicket { get; set; }
 }
 
 
 public class AnexoModel
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public required byte[] Arquivo { get; set; }
     public required string Tipo { get; set; }
-    public int IdMensagem { get; set; }
+    public long IdMensagem { get; set; }
     public DateTime DataHoraCriado { get; set; }
 }
